+++ ChangeLog.md
@@ -23,6 +23,8 @@
 * Bug Fix: GetValues<object> on empty element may throw ArgumentOutOfRangeException (#720)
 * Serilog for .NET Standard 1.3/.NET Core (#772)
 * Bug Fix : DicomServer does not close TcpClient/socket after client closed connection (#795)
+* Bug Fix : JsonDicomConverter should decode BulkDataURI element(#796)
+* Bug Fix: fix deserialization of Json when the VR-property is not on first position (#730) 
 * DicomClient uses StrongBox to reduce memory consumption (#794)
 * Bug Fix: C-STORE request may hang. (#792)